Split assert code into distinct module
